refactor: move patch closer to tiptap context
