+++ frontend/src/stores/element.js
@@ -19,7 +19,7 @@ import { presentationId } from './presentation'
 import { getUpdatedIdAfterConnections } from './transition'
 
 import { generateHTML } from '@tiptap/core'
-import { extensions } from '@/stores/tiptapSetup'
+import { extensions, patchEmptyParagraphs } from '@/stores/tiptapSetup'
 
 const activeElementIds = ref([])
 const focusElementId = ref(null)
@@ -469,43 +469,6 @@ const addFixedWidthToElement = (deltaWidth) => {
 
 const { initTextEditor, activeEditor } = useTextEditor()
 let editorOldText = ''
-
-const patchEmptyParagraphs = (htmlString) => {
-	let didUpdate = false
-
-	const parser = new DOMParser()
-	const doc = parser.parseFromString(htmlString, 'text/html')
-
-	const allParagraphs = Array.from(doc.body.querySelectorAll('p'))
-	let prevSpanStyles = null
-
-	allParagraphs.forEach((p, index) => {
-		const isEmpty = p.textContent.trim() === ''
-		const firstSpan = p.querySelector('span')
-
-		if (!isEmpty && firstSpan) {
-			prevSpanStyles = firstSpan.getAttribute('style') || ''
-			return
-		}
-
-		if (isEmpty && prevSpanStyles) {
-			p.innerHTML = ''
-
-			if (!didUpdate) didUpdate = true
-
-			const span = doc.createElement('span')
-			span.setAttribute('style', prevSpanStyles)
-			span.innerHTML = '\u200B'
-
-			p.appendChild(span)
-		}
-	})
-
-	return {
-		wasUpdated: didUpdate,
-		updatedHTML: didUpdate ? doc.body.innerHTML : htmlString,
-	}
-}
 
 const getEditorHTML = () => {
 	const html = activeEditor.value.getHTML()
